# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SilentRelay API
  description: |
    Secure end-to-end encrypted messaging API for SilentRelay.
    
    ## Authentication
    Most endpoints require JWT authentication via `Authorization: Bearer <token>` header.
    
    ## Rate Limiting
    - SMS endpoints: 5 requests/minute/IP
    - Auth endpoints: 10 requests/minute/IP  
    - Protected endpoints: 100 requests/minute/user
    
    ## Security
    All communications use end-to-end encryption. The server never has access to message content.
  version: 1.0.0
  contact:
    name: SilentRelay Support
    url: https://silentrelay.com.au
  license:
    name: Proprietary
    url: https://silentrelay.com.au/terms
servers:
  - url: https://silentrelay.com.au/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or registration
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone_number:
          type: string
          example: "+1234567890"
        username:
          type: string
          example: "john_doe"
        display_name:
          type: string
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        created_at:
          type: string
          format: date-time
...
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        creator_id:
          type: string
          format: uuid
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
    PrivacySettings:
      type: object
      properties:
        profile_photo:
          type: string
          enum: [everyone, contacts, nobody]
        last_seen:
          type: string
          enum: [everyone, contacts, nobody]
        read_receipts:
          type: boolean
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user
      description: Returns the authenticated user's profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

    put:
      tags:
        - Users
      summary: Update current user
      description: Updates the authenticated user's profile
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                username:
                  type: string
                avatar_url:
                  type: string
                  format: uri
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      tags:
        - Users
      summary: Delete current user
      description: Permanently deletes the user account and all associated data
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted
        '401':
          description: Unauthorized
...
  /users/{userId}/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Retrieves a user's public profile
      operationId: getUserProfile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /messages:
    get:
      tags:
        - Messages
      summary: Get messages
      description: Retrieves pending encrypted messages for the authenticated user
      operationId: getMessages
      security:
        - bearerAuth: []
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Return messages since this timestamp
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        '200':
          description: List of encrypted messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        sender_id:
                          type: string
                          format: uuid
                        encrypted_content:
                          type: string
                          format: byte
                        timestamp:
                          type: string
                          format: date-time
  /messages/{messageId}/status:
    put:
      tags:
        - Messages
      summary: Update message status
      description: Updates the delivery/read status of a message
      operationId: updateMessageStatus
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [delivered, read]
      responses:
        '200':
          description: Status updated

  # =====================
  # GROUPS
  # =====================
  /groups:
    post:
      tags:
        - Groups
      summary: Create group
      description: Creates a new group chat
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - members
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                members:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        '200':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
  /groups/{groupId}:
    get:
      tags:
        - Groups
      summary: Get group
      description: Retrieves group details
      operationId: getGroup
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '403':
          description: Not a group member
        '404':
          description: Group not found
  /groups/{groupId}/members:
    post:
      tags:
        - Groups
      summary: Add group member
      description: Adds a new member to the group
      operationId: addGroupMember
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Member added
  /groups/{groupId}/members/{userId}:
    delete:
      tags:
        - Groups
      summary: Remove group member
      description: Removes a member from the group
      operationId: removeGroupMember
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed

  # =====================
  # MEDIA
  # =====================
...
  /privacy:
    get:
      tags:
        - Privacy
      summary: Get privacy settings
      operationId: getPrivacySettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Privacy settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacySettings'

    post:
      tags:
        - Privacy
      summary: Update privacy setting
      operationId: updatePrivacySetting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                setting:
                  type: string
                  enum: [profile_photo, last_seen, read_receipts]
                value:
                  type: string
      responses:
        '200':
          description: Setting updated
...